third_party/blink/web_tests/wpt_internal/app-history/navigate-event/respondWith-resolve.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
async_test(t => {
  appHistory.onnavigate = e => {
    e.respondWith(new Promise(resolve => t.step_timeout(() => {
      resolve();
      t.step_timeout(t.step_func_done(() => assert_equals(location.hash, "#1")), 0);
    }, 0)));
  };

  location.href = "#1";
  assert_equals(location.hash, "#1");
}, "AppHistoryNavigateEvent.respondWith should proceed if the given promise resolves");
</script>
